+++ plugins/monolith/src/ortac_monolith.ml
@@ -132,12 +132,9 @@ let standalone module_name env s =
   :: specs
 
 let generate path fmt =
-  let module_name = Ortac_core.Utils.module_name_of_path path in
-  Gospel.Parser_frontend.parse_ocaml_gospel path
-  |> Ortac_core.Utils.type_check [] path
-  |> fun (env, sigs) ->
-  assert (List.length env = 1);
-  standalone module_name (List.hd env) sigs
+  let open Ortac_core.Utils in
+  let { module_name; namespace; ast } = check path in
+  standalone module_name namespace ast
   |> Fmt.pf fmt "%a@." Ppxlib_ast.Pprintast.structure;
   W.report ()
 
@@ -159,7 +156,7 @@ end = struct
     Cmd.info "monolith"
       ~doc:"Generate Monolith test file according to Gospel specifications."
 
-  let term = Term.(const main $ ocaml_file $ output_file $ setup_log)
+  let term = Term.(const main $ input_file $ output_file $ setup_log)
   let cmd = Cmd.v info term
 end